pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-options-agent"
version = "0.1.0"
description = "AI Options Trading Agent — prediction markets, sentiment, RL"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Developer Team" }]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "requests>=2.28",
    "httpx>=0.24",
    "psycopg2-binary>=2.9",
    "py-clob-client>=0.14",
    "python-dotenv>=1.0",
    "cryptography>=41.0",
    "python-dateutil>=2.8",
    "yfinance>=0.2",
    "vaderSentiment>=3.3",
    "transformers>=4.30,<4.46",
    "torch>=2.0",
    "praw>=7.0",
    "gymnasium>=0.29",
    "numpy>=1.24",
    "stable-baselines3>=2.0",
    "scipy>=1.9",
    "matplotlib>=3.5",
    "streamlit>=1.30",
    "pandas>=2.0",
    "plotly>=5.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "respx>=0.20",
]
# Optional: for computing Black–Scholes Greeks when provider omits them (legacy setup.py; install setuptools first if needed)
greeks = [
    "py_vollib>=1.0",
]

[project.scripts]
backfill = "scripts.backfill:main"
incremental_update = "scripts.incremental_update:main"
build_pm_features = "scripts.build_pm_features:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
